Refactor: 실전모드 SelectSeat 한 파일로 통일 및 SecureModal 위치 변경
