+++ src/pages/challengeMode/selectSeat/SelectSeatChallangeMode.jsx
@@ -1,0 +1,73 @@
+import React, { useEffect, useState } from "react";
+import styled from "styled-components";
+import SecureModal from "../../../components/modal/SecureModal";
+import SeatChart from "../../../components/seatChart/SeatChart1";
+import SeatSection from "../../../components/seatSection/SeatSection";
+import { useAtomValue, useSetAtom } from "jotai";
+import {
+  isSectionSelectedAtom,
+  progressAtom,
+  themeSiteAtom
+} from "../../../store/atom";
+import SeatInfo from "../../../components/seatInfo/SeatInfo";
+import ErrorTooltip from "../../../components/tooltip/ErrorTooltip";
+
+const SelectSeatContainer = styled.div`
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  gap: 80px;
+`;
+
+const SeatInfoContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+const TooltipText = styled.div`
+  width: 380px;
+  font-family: PretendardM;
+`;
+
+const SelectSeatChallangeMode = () => {
+  const [isModalOpen, setIsModalOpen] = useState(true);
+  const isSectionSelected = useAtomValue(isSectionSelectedAtom);
+  const setProgress = useSetAtom(progressAtom);
+  const themeSite = useAtomValue(themeSiteAtom);
+
+  useEffect(() => {
+    setProgress(2);
+  }, []);
+
+  const closeModal = () => {
+    setIsModalOpen(false);
+  };
+
+  return (
+    <SelectSeatContainer>
+      {isModalOpen && <SecureModal onClick={closeModal} />}
+      {isSectionSelected ? (
+        <SeatChart></SeatChart>
+      ) : (
+        <SeatSection></SeatSection>
+      )}
+
+      <SeatInfoContainer>
+        {/* 에러 툴팁은 멜론티켓 페이지에서만 렌더링 */}
+        {themeSite === "melonticket" && (
+          <ErrorTooltip
+            contents={
+              <TooltipText>
+                좌석 선택 이후 5분내로 결제가 완료되지 않을시 <br />
+                선택하신 좌석의 선점기회를 잃게됩니다
+              </TooltipText>
+            }
+          ></ErrorTooltip>
+        )}
+        <SeatInfo></SeatInfo>
+      </SeatInfoContainer>
+    </SelectSeatContainer>
+  );
+};
+export default SelectSeatChallangeMode;
